Removed dedicated dynvar slot
Removed the slot specifically created to hold the culling distance driver per user. The variable is named so there's not much reason to give it it's own slot. The global variable currently has a slot with a tag so it may still be needed but it may be removed in the future.